Change to use a kwarg dict for cutout config
Instead of duplicating the cutout config keyword list, just pass dicts
that can be filled with the desired overrides.
Use two cutout plotters, so that they can be configured differently.
Config for the plotter needs the chunk size as well; default it to None, for
easier user access to the plots.