Revert 4 commits
This reverts to before commit 80930f922f34449d1d8ed1eaa99d24ce10f0cbe1.

This was an unintentional force push to main; I have reset the
permissions to prevent this in the future. These 4 commits will become
part of tickets/DM-40387 and will be in an ordinary PR there.