Change sqlite pool parameters
Default pool size (5) was causing timeouts in tests with limit=5;
increasing significantly allows test to pass, and decreasing timeout lets
things fail sooner (5 seconds should be plenty for our sqlite transactions).
The goal is to keep the sqlite connection open through the entire set
of transactions, so that we are not repeatedly re-scanning the master
table.